+++ nara/extra/Datetime/today_time.py
@@ -1,194 +1,141 @@
-import asyncio
-import re
-import aiohttp
-from bs4 import BeautifulSoup
-from datetime import datetime,time
+from datetime import datetime
+from typing import Union
 
 import requests
 
-async def fetch_url(session, url):
-    async with session.get(url, headers={"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"}) as response:
-        if response.status == 200:return await response.text()
-        else: return False
-async def async_CurrentDateTime(current_timestamp=False,Current_Date=False,encode=False,Country='India')->str|datetime:
-    url = f"https://time.is/{Country}"
-    async with aiohttp.ClientSession() as session:
-        html = await fetch_url(session, url)
+
+class CurrentDateTime:
+    """
+    A class to retrieve current date, time, and country codes from a remote API.
+
+    Methods
+    -------
+    country_codes() -> dict:
+        Retrieve a dictionary of country codes from a remote API endpoint.
+
+    time(country_code: str, encode: bool = False) -> Union[dict, dict[str, datetime]]:
+        Retrieve current time information for a specified country code.
+
+    date(country_code: str, encode: bool = False) -> Union[str, datetime]:
+        Retrieve current date information for a specified country code.
+    """
+    def country_codes()->dict:
+        """
+        Retrieve a dictionary of country codes from a remote API endpoint.
+
+        Returns
+        -------
+        dict
+            A dictionary where keys are country names and values are their corresponding codes.
+        """
+        headers= {
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"
+        }
+        html_response = requests.get(f"https://time-epbx.onrender.com/", headers=headers)
+        return html_response.json().get("country_codes")
         
-        if not html:raise Exception("Invalid Country name or error while fetching time date")
-        soup = BeautifulSoup(html, 'html.parser')
+    def time(country_code:str,encode=False) -> dict|datetime:
+        """
+        Retrieve current time information for a specified country code.
 
-        if current_timestamp and Current_Date:
-            try:
-                clock_element = soup.find('time', id='clock')
-                if clock_element:
-                    time_obj = clock_element.text.strip()
-                    if encode:
-                        try:time_obj = datetime.strptime(time_obj, "%H:%M:%S").time()
-                        except:
-                            try:
-                                time_obj = datetime.strptime(time_obj, "%I:%M:%S%p")
-                            except:time_obj = datetime.strptime(time_obj,"%I:%M:%S %p")
-                
-                date_element = soup.find('div', class_='clockdate')
-                if date_element:
-                    date_info = date_element.text.strip()
-                    
-                    # Extract and format the date
-                    date_parts = date_info.split(',')
-                    current_date_obj = date_parts[1].strip() + ' ' + date_parts[2].strip()
-                    # print("Formatted Date:", current_date_obj)
-                    if encode:
-                        try:
-                            current_date_obj = datetime.strptime(current_date_obj, r'%B %d %Y').date()
-                        except:
-                            current_date_obj = datetime.strptime(current_date_obj, r'%d %B %Y').date()
-                    
-                    return time_obj, current_date_obj
-            except:return None, None
-        elif current_timestamp:
-            # Extract the time
-            clock_element = soup.find('time', id='clock')
-            if clock_element:
-                time_obj = clock_element.text.strip()
+        Parameters
+        ----------
+        country_code : str
+            The country code for which time information is requested.
+        encode : bool, optional
+            Whether to return encoded datetime objects (`True`) or time strings (`False`). Default is `False`.
+
+        Returns
+        -------
+        Union[dict, dict[str, datetime]]
+            If `encode` is `True`, returns a dictionary with 12-hour and 24-hour time as `datetime` objects.
+            If `encode` is `False`, returns a dictionary with 12-hour and 24-hour time as strings.
+
+        Notes
+        -----
+        This function sends a GET request to an external API endpoint to fetch current time information
+        for the specified country code. It expects the endpoint to return JSON with time data.
+
+        Example
+        -------
+        >>> CurrentDateTime.time("IN")
+        {'12_hour': '09:30:00 PM', '24_hour': '21:30:00'}
+        >>> CurrentDateTime.time("IN", encode=True)
+        {'12_hour': datetime.datetime(2024, 7, 8, 21, 30), '24_hour': datetime.datetime(2024, 7, 8, 21, 30)}
+        """
+        headers= {
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"
+    }
+
+        # Define the format of the input strings
+        time_12_hour_format = r'%I:%M:%S %p'
+        time_24_hour_format = r'%H:%M:%S'
+        html_response = requests.get(f"https://time-epbx.onrender.com/?country={country_code}", headers=headers)
+        if html_response.status_code!=200:return html_response.json()
+        elif html_response.status_code == 200:
+            html_response:dict = html_response.json()
+            if html_response.get("error") is None:
+                time_12_hour = ' '.join(html_response.get("time").get("12_hour").split()[:2])
+                time_24_hour = ' '.join(html_response.get("time").get("24_hour").split()[:1])
                 if encode:
-                    try:time_obj = datetime.strptime(time_obj, "%H:%M:%S").time()
-                    except:
-                        time_obj = re.sub(r"(\d{2}:\d{2}:\d{2})([APM]+)", r"\1 \2", time_obj)
-                        
-                        try:
-                            time_obj = datetime.strptime(time_obj, "%I:%M:%S%p")
-                        except:time_obj = datetime.strptime(time_obj,"%I:%M:%S %p")
-
-                # print(time_obj)
-
-                return time_obj
+                    print(f"{time_24_hour = },{time_12_hour = }")
+                    time_12_hour_obj = datetime.strptime(time_12_hour, time_12_hour_format)
+                    time_24_hour_obj = datetime.strptime(time_24_hour, time_24_hour_format)
+                    return {"12_hour": time_12_hour_obj, "24_hour": time_24_hour_obj}
+                else:
+                    return {"12_hour":time_12_hour,"24_hour":time_24_hour}
             else:
-                return 
+                return html_response.get("error")
+    def date(country_code:str,encode=False) -> Union[str, datetime]:
+        """
+        Retrieve current date information for a specified country code.
 
-        ################### CURRENT DATE #####################
-        elif Current_Date:        # Extract the date
-            date_element = soup.find('div', class_='clockdate')
-            if date_element:
-                date_info = date_element.text.strip()
-                
-                # Extract and format the date
-                date_parts = date_info.split(',')
-                current_date_obj = date_parts[1].strip() + ' ' + date_parts[2].strip()
-                # print("Formatted Date:", current_date_obj)
-                if encode:
-                    try:
-                        current_date_obj = datetime.strptime(current_date_obj, r'%B %d %Y').date()
-                    except:
-                        current_date_obj = datetime.strptime(current_date_obj, r'%d %B %Y').date()
+        Parameters
+        ----------
+        country_code : str
+            The country code for which date information is requested.
+        encode : bool, optional
+            Whether to return the date as a `datetime` object (`True`) or as a string (`False`). Default is `False`.
 
-                return current_date_obj
-            
+        Returns
+        -------
+        Union[str, datetime]
+            If `encode` is `True`, returns the date as a `datetime` object.
+            If `encode` is `False`, returns the date as a string.
+
+        Notes
+        -----
+        This function sends a GET request to an external API endpoint to fetch current date information
+        for the specified country code. It expects the endpoint to return JSON with date data.
+
+        Example
+        -------
+        >>> CurrentDateTime.date("IN")
+        '2024-07-08'
+        >>> CurrentDateTime.date("IN", encode=True)
+        datetime.datetime(2024, 7, 8, 0, 0)
+        """
+        headers= {
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"
+    }
+
+        # Define the format of the input strings
+        html_response = requests.get(f"https://time-epbx.onrender.com/?country={country_code}", headers=headers)
+        if html_response.status_code!=200:return html_response.json()
+        elif html_response.status_code == 200:
+            html_response:dict = html_response.json()
+            if html_response.get("error") is None:
+                if encode:                    
+                    return datetime.strptime(html_response.get("date"), "%Y-%m-%d")
+                else:
+                    return html_response.get("date")
             else:
-                return None
-
-def CurrentDateTime(current_timestamp=False,Current_Date=False,encode=False,Country='India')->str|datetime:
-    headers= {
-    "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
-    "accept-language": "en-US,en;q=0.9",
-    "cache-control": "max-age=0",
-    "priority": "u=0, i",
-    "sec-ch-ua": "\"Not/A)Brand\";v=\"8\", \"Chromium\";v=\"126\", \"Google Chrome\";v=\"126\"",
-    "sec-ch-ua-mobile": "?0",
-    "sec-ch-ua-platform": "\"Windows\"",
-    "sec-fetch-dest": "document",
-    "sec-fetch-mode": "navigate",
-    "sec-fetch-site": "none",
-    "sec-fetch-user": "?1",
-    "upgrade-insecure-requests": "1",
-    "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36"
-  }
-    html = requests.get(f"https://time.is/{Country}", headers=headers)
-    if html.status_code!=200:raise Exception("Invalid Country name or error while fetching time date",html.status_code)
-    elif html.status_code == 200:
-        html = html.content
-        soup = BeautifulSoup(html, 'html.parser')
-        if current_timestamp and Current_Date:
-            try:
-                clock_element = soup.find('time', id='clock')
-                if clock_element:
-                    time_obj = clock_element.text.strip()
-                    if encode:
-                        try:time_obj = datetime.strptime(time_obj, "%H:%M:%S").time()
-                        except:
-                            try:
-                                time_obj = datetime.strptime(time_obj, "%I:%M:%S%p")
-                            except:time_obj = datetime.strptime(time_obj,"%I:%M:%S %p")
-                
-                date_element = soup.find('div', class_='clockdate')
-                if date_element:
-                    date_info = date_element.text.strip()
-                    
-                    # Extract and format the date
-                    date_parts = date_info.split(',')
-                    current_date_obj = date_parts[1].strip() + ' ' + date_parts[2].strip()
-                    # print("Formatted Date:", current_date_obj)
-                    if encode:
-                        try:
-                            current_date_obj = datetime.strptime(current_date_obj, r'%B %d %Y').date()
-                        except:
-                            current_date_obj = datetime.strptime(current_date_obj, r'%d %B %Y').date()
-                    
-                    return time_obj, current_date_obj
-            except:return None, None
-        elif current_timestamp:
-            # Extract the time
-            clock_element = soup.find('time', id='clock')
-            if clock_element:
-                time_obj = clock_element.text.strip()
-                if encode:
-                    try:time_obj = datetime.strptime(time_obj, "%H:%M:%S").time()
-                    except:
-                        time_obj = re.sub(r"(\d{2}:\d{2}:\d{2})([APM]+)", r"\1 \2", time_obj)
-                        
-                        try:
-                            time_obj = datetime.strptime(time_obj, "%I:%M:%S%p")
-                        except:time_obj = datetime.strptime(time_obj,"%I:%M:%S %p")
-
-                # print(time_obj)
-
-                return time_obj
-            else:
-                return 
-
-        ################### CURRENT DATE #####################
-        elif Current_Date:        # Extract the date
-            date_element = soup.find('div', class_='clockdate')
-            if date_element:
-                date_info = date_element.text.strip()
-                
-                # Extract and format the date
-                date_parts = date_info.split(',')
-                current_date_obj = date_parts[1].strip() + ' ' + date_parts[2].strip()
-                # print("Formatted Date:", current_date_obj)
-                if encode:
-                    try:
-                        current_date_obj = datetime.strptime(current_date_obj, r'%B %d %Y').date()
-                    except:
-                        current_date_obj = datetime.strptime(current_date_obj, r'%d %B %Y').date()
-
-                return current_date_obj
-            
-            else:
-                return None
-    
-# Example usage
-async def main():
-    current_time,current_date = await async_CurrentDateTime(current_timestamp=True,Current_Date=True)
-    print("current_time = ",current_time,type(current_date),"current_date = ",current_date,type(current_date))
+                return html_response.get("error")
 
 if __name__ == "__main__":
-    # Run the async function
-    asyncio.run(main())
-    a =CurrentDateTime(current_timestamp=True,encode=True)
-    # a = datetime.strptime("00:01:12", "%H:%M:%S").time()
-    print(a,type(a))
-    if a and a.strftime("%p") == "AM" and (a.hour == 00 and a.minute in [0,1,2,3]):
-        print(True)
-    else:
-        print(False)
+    datetime_obj = CurrentDateTime
+    print(datetime_obj.time("in"))
+    print(datetime_obj.time("in",encode=True))
+    print(datetime_obj.date("in",encode=True))
+    print(datetime_obj.date("in"))
+    print(datetime_obj.country_codes())