+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+### 0.3 - 09-7-2024
+>  ## Added new modules  
+> - async_thred_task :  Runs an asynchronous function in a separate thread
+> - clear_pycache_directories : Removes all __pycache__ directories recursively.
+> - display_file_structure : Recursively builds and prints a tree with directory contents.
+> - retry : Retries a function upon failure with specified retry logic.
+> - repeat_forever : Runs a function in an infinite loop with optional sleep intervals.
+> - run_multiple_times : Runs an asynchronous function multiple times concurrently.
+> - RandomDataGenerator : generate random date like name,id,password,location,indian name,etc.
+
+
+* fix tele_db bugs, and configuration 
+* Update Current Date and Time module
+* Update TimeIt function - now you can use it with async and sync function
+
 ### 0.2.3 - 20-6-2024
 * fix tele_db bugs, currentTime and configuration